+++ datumaro/plugins/ade20k2020_format.py
@@ -179,5 +179,7 @@ class Ade20k2020Importer(Importer):
         for i in range(5):
             for i in glob.iglob(osp.join(path, *('*' * i))):
                     if osp.splitext(i)[1].lower() in IMAGE_EXTENSIONS:
-                        return [{'url': path, 'format': 'ade20k2020'}]
+                        return [{
+                            'url': path, 'format': Ade20k2020Extractor.NAME,
+                        }]
         return []